improve age-structure-adjustment
Previously, the "Total" category was not calculated as sum of deaths in each category adjusted for the size of that category, but as its own category, hence the age-normalised values in Total did not reflect the changes in age composition inside the Total from year to year. This is now fixed.